utils working
